refactor: FollowCardList 분리
